Johannes corrections 2
