+++ tenpy/models/bose_hubbard_chain.py
@@ -1,9 +1,9 @@
-"""The contents of this module have been moved to :mod:`tenpy.models.bose_hubbard`.
+"""The contents of this module have been moved to :mod:`tenpy.models.hubbard`.
 
 This module is just around for backwards compatibility."""
-# Copyright 2018 TeNPy Developers
+# Copyright 2019 TeNPy Developers
 
-from .bose_hubbard import BoseHubbardModel, BoseHubbardChain
+from .hubbard import BoseHubbardModel, BoseHubbardChain
 
 import warnings
 
@@ -13,6 +13,6 @@ msg = """BUGFIX
 * The Hamiltonian of the `BoseHubbardModel` was previously implemented with $ U n_i^2 $ as interaction term,
 * but documented as $ U n_i (n_i-1)$. Now it is implemented as the latter as well.
 ***********
-To avoid this warning, simply import the model class from `tenpy.models.bose_hubbard` instead of `tenpy.models.bose_hubbard_chain`."""
+To avoid this warning, simply import the model class from `tenpy.models.hubbard` instead of `tenpy.models.bose_hubbard_chain`."""
 warnings.warn(msg)
 warnings.warn("The module `tenpy.models.bose_hubbard_chain` is deprecated now.", FutureWarning)
