+++ tenpy/models/lattice.py
@@ -85,10 +85,6 @@ class Lattice:
         Same as `nearest_neighbors`, but for the next-nearest neighbors.
     next_next_nearest_neighbors : ``None`` | list of ``(u1, u2, dx)``
         Same as `nearest_neighbors`, but for the next-next-nearest neighbors.
-    fourth_nearest_neighbors : ``None`` | list of ``(u1, u2, dx)``
-        Same as `nearest_neighbors`, but for the fourth-nearest neighbors. (Only implemented for Honeycomb).
-    fifth_nearest_neighbors : ``None`` | list of ``(u1, u2, dx)``
-        Same as `nearest_neighbors`, but for the fifth-nearest neighbors. (Only implemented for Honeycomb).
 
     Attributes
     ----------
@@ -132,10 +128,6 @@ class Lattice:
         Same as :attr:`nearest_neighbors`, but for the next-nearest neighbors.
     next_next_nearest_neighbors : ``None`` | list of ``(u1, u2, dx)``
         Same as :attr:`nearest_neighbors`, but for the next-next-nearest neighbors.
-    fourth_nearest_neighbors : ``None`` | list of ``(u1, u2, dx)``
-        Same as :attr:`nearest_neighbors`, but for the fourth-nearest neighbors. (Only implemented for Honeycomb).
-    fifth_nearest_neighbors : ``None`` | list of ``(u1, u2, dx)``
-        Same as :attr:`nearest_neighbors`, but for the fifth-nearest neighbors. (Only implemented for Honeycomb).
     _order : ndarray (N_sites, dim+1)
         The place where :attr:`order` is stored.
     _strides : ndarray (dim, )
@@ -160,9 +152,8 @@ class Lattice:
                  positions=None,
                  nearest_neighbors=None,
                  next_nearest_neighbors=None,
-                 next_next_nearest_neighbors=None,
-                 fourth_nearest_neighbors=None,
-                 fifth_nearest_neighbors=None):
+                 next_next_nearest_neighbors=None
+                 ):
         self.Ls = tuple([int(L) for L in Ls])
         self.unit_cell = list(unit_cell)
         self.N_cells = int(np.prod(self.Ls))
@@ -189,8 +180,6 @@ class Lattice:
         self.nearest_neighbors = nearest_neighbors
         self.next_nearest_neighbors = next_nearest_neighbors
         self.next_next_nearest_neighbors = next_next_nearest_neighbors
-        self.fourth_nearest_neighbors = fourth_nearest_neighbors
-        self.fifth_nearest_neighbors = fifth_nearest_neighbors
         self.test_sanity()  # check consistency
 
     def test_sanity(self):
@@ -1061,12 +1050,6 @@ class Chain(SimpleLattice):
         kwargs.setdefault('next_next_nearest_neighbors', [(0, 0, np.array([
             3,
         ]))])
-        kwargs.setdefault('fourth_nearest_neighbors', [(0, 0, np.array([
-            4,
-        ]))])
-        kwargs.setdefault('fifth_nearest_neighbors', [(0, 0, np.array([
-            5,
-        ]))])
         # and otherwise default values.
         SimpleLattice.__init__(self, [L], site, **kwargs)
 
@@ -1182,7 +1165,8 @@ class Honeycomb(Lattice):
         If only a single :class:`~tenpy.networks.site.Site` is given, it is used for both sites.
     **kwargs :
         Additional keyword arguments given to the :class:`Lattice`.
-        `basis`, `pos` and `[[next_]next_]nearest_neighbors` (etc.) are set accordingly.
+        `basis`, `pos` and `[[next_]next_]nearest_neighbors` are set accordingly.
+        Additionally, for the Honeycomb lattice fourth_nearest_neighbors and fifth_nearest_neighbors are implemented.
     """
     dim = 2
 
@@ -1197,15 +1181,13 @@ class Honeycomb(Lattice):
         nNN = [(0, 0, np.array([1, 0])), (0, 0, np.array([0, 1])), (0, 0, np.array([1, -1])),
                (1, 1, np.array([1, 0])), (1, 1, np.array([0, 1])), (1, 1, np.array([1, -1]))]
         nnNN = [(1, 0, np.array([1, 1])), (0, 1, np.array([-1, 1])), (0, 1, np.array([1, -1]))]
-        fourthNN = [(0, 1, np.array([0, 1])), (0, 1, np.array([1, 0])), (0, 1, np.array([1, -2])),
-                    (0, 1, np.array([0, -2])), (0, 1, np.array([-2, 0])), (0, 1, np.array([-2, 1]))]
-        fifthNN = [(0, 0, np.array([1, 1])), (0, 0, np.array([2, -1])), (0, 0, np.array([1, -2])),
-                   (0, 0, np.array([-1, -1])), (0, 0, np.array([-2, 1])), (0, 0, np.array([-1, 2]))]
+        self.fourth_nearest_neighbors = [(0, 1, np.array([0, 1])), (0, 1, np.array([1, 0])), (0, 1, np.array([1, -2])),
+                                         (0, 1, np.array([0, -2])), (0, 1, np.array([-2, 0])), (0, 1, np.array([-2, 1]))]
+        self.fifth_nearest_neighbors = [(0, 0, np.array([1, 1])), (0, 0, np.array([2, -1])), (0, 0, np.array([1, -2])),
+                                        (0, 0, np.array([-1, -1])), (0, 0, np.array([-2, 1])), (0, 0, np.array([-1, 2]))]
         kwargs.setdefault('nearest_neighbors', NN)
         kwargs.setdefault('next_nearest_neighbors', nNN)
         kwargs.setdefault('next_next_nearest_neighbors', nnNN)
-        kwargs.setdefault('fourth_nearest_neighbors', fourthNN)
-        kwargs.setdefault('fifth_nearest_neighbors', fifthNN)
         Lattice.__init__(self, [Lx, Ly], sites, **kwargs)
 
     def ordering(self, order):
